requirements.txt
Django==3.2.12
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Pillow==9.0.1
whitenoise==6.0.0